api/ImageSimilarity/deployment/requirements_gpu.txt
numpy==1.14.3
pandas==0.22.0
Keras==2.1.6
Pillow==5.3.0
azure_common==1.1.16
scikit_learn==0.19.1
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
flask-cors
annoy
tensorflow-gpu>1.5
